Move/renamve InventoryGroupPage to GroupPage
